requirements.txt
attrs==21.4.0
Brotli==1.0.9
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dash==2.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.3
gunicorn=20.1.0
hpsklearn==0.0.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
kiwisolver==1.4.2
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mistune==2.0.2
numba==0.55.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
Platypus==1.0
Platypus-Opt==1.0.4
plotly==5.8.0
ply==3.11
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.1.2
